Workflow to verify the OSX latest version
Workflow triggers when a push is made to main branch which is common with a version update. Installs Besu on OSX runner
and check the installed version is same as the version configured in besu.rb file. This verifies that published
version is the latest version installed on OSX using Homebrew

Signed-off-by: Chaminda Divitotawela <cdivitotawela@gmail.com>